requirements.txt
python-docx
python-dotenv
google-generativeai